chore(#289): retire ADR-006's unadopted snapshot capsule, keep the part in use
`VersionedSnapshot`, `StateSource`, `Completeness`, `etag` and `cacheAgeMillis`
were scaffolding for a `RefreshCoordinator` increment that never arrived. #675
shipped the behaviour that coordinator existed to provide - single-flight,
cancellation, latest-value-wins - without any of it, and the remaining pieces of
its proposed API do not survive contact with what was learned since:

  priority               withdrawn as unsound on #664: rank is not freshness
  expectedProjectEpoch   the compare-and-swap already refuses stale writes
  per-section refresh    the one capability not otherwise covered, and #668
                         measured the whole cycle at 16.22s worst case against a
                         25s deadline, so it is an optimisation nothing needs

`docs/adr/README.md` had already recorded the capsule as unadopted with "only
CacheSectionID is live". A per-symbol census confirmed it exactly:

  CacheSectionID      12 production call sites
  StateSource          0
  Completeness         0
  VersionedSnapshot    0  (tests only)

My first count said `Completeness` had one production use. It was the word
"Completeness" starting a sentence in a doc comment - grep counting prose, not
code. That is also why the file is renamed rather than deleted: I had written a
next action to delete it outright, which would have broken twelve call sites.

The tests go with the types. They passed while proving nothing about the
product, which is the shape of coverage that reads as real in a later census -
the trap this issue already fell into once, when an `ADR-0XX` string search
reported modules as "not started" that were named for what they do.

The ADR honesty note is updated to say retired rather than deferred, so the
prose and the code agree.
