+++ docs/adr/README.md
@@ -80,7 +80,7 @@ Kernel increments merged to `main`. State-changing pilot paths stay behind their
 - **Runtime-wired (flag-gated):** the resource read-path cache envelope now carries `project_epoch` (from the ADR-002 `TargetRegistry`), `section_revision` (a per-`CacheSectionID` monotonic counter added to `StateCache`), and `etag` (a pure FNV-1a fingerprint of the body) on the `logic://tracks` / `logic://transport` / `logic://mixer` / `logic://project` resources — **only when `FeatureFlags.adr006VersionedCache` is on**. With the flag off the envelope is **byte-identical** to before (proven by dedicated byte-identity tests plus the pre-existing envelope tests still passing), so this adds zero runtime behavior by default.
 - Still deferred: the `RefreshCoordinator` (single-flight refresh + backpressure) and the 30-minute memory-soak / large-project benchmarks from #289's acceptance criteria — those need a live qualification environment and are not claimed here.
 - Deterministic tests only (no runtime path touched): field preservation, etag stability, `cacheAgeMillis` monotonicity + clock-skew guard, `Codable` round-trips, section completeness, flag-default-off.
-- **Honesty notes:** the `VersionedSnapshot` capsule (`StateSource`/`Completeness`/`cacheAgeMillis`) is unadopted scaffolding — the live envelope path uses its own `ResourceHandlers` etag/extras helpers; adoption is deferred to the `RefreshCoordinator` increment (only `CacheSectionID` is live). The epoch flag-coupling defect (versioned cache on + target refs off froze `project_epoch` at 0 across project switches) is fixed: lifecycle invalidation now bumps the epoch when **either** consumer flag is active, with a dispatcher-path regression test.
+- **Honesty notes:** the `VersionedSnapshot` capsule (`StateSource`/`Completeness`/`cacheAgeMillis`) has been **retired**. It was scaffolding awaiting a `RefreshCoordinator` increment that never arrived, and #675 shipped the behaviour it existed to support — single-flight, cancellation, latest-value-wins — without it. A per-symbol census found zero production callers for all of it and 12 for `CacheSectionID`, which is what remains (`Sources/LogicProMCP/State/CacheSectionID.swift`). The live envelope path continues to derive its own etag and cache age in `ResourceHandlers`. The epoch flag-coupling defect (versioned cache on + target refs off froze `project_epoch` at 0 across project switches) is fixed: lifecycle invalidation now bumps the epoch when **either** consumer flag is active, with a dispatcher-path regression test.
 
 ### ADR-004 — Verified Mutation Saga (`In Implementation`)
 - The pure in-memory `actor MutationSaga` remains behind `FeatureFlags.adr004MutationSaga` (default off). The public `logic_system` surface now exposes `saga_preflight`, `saga_execute`, `saga_status`, and `saga_cancel` through the six-operation production executor and a bounded, session-only journal.
